refactor: ProductScheduleStockService 제거 및 재고 차감 로직을 ProductScheduleCommandService로 통합
