+++ src/main/java/com/yeoljeong/tripmate/product/application/service/query/ProductSearchService.java
@@ -11,15 +11,15 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
 import org.springframework.stereotype.Service;
 
-// 날짜 기준으로 판매 가능한 상품 조회
-//ex ) 5월 28일 -> 박물관 입장권 , 성당 입장권 , 낙타 투어  ..
+
 @Service
 @RequiredArgsConstructor
 public class ProductSearchService {
 
   private final ProductScheduleRepository scheduleRepository;
 
-  // 날짜 기준으로 예약 가능한 상품 조회
+  // 날짜 기준으로 판매 가능한 상품 조회
+  //ex ) 5월 28일 -> 박물관 입장권 , 성당 입장권 , 낙타 투어  ..
   public Slice<ProductScheduleInfoResult> getAvailableProducts(LocalDate date, Pageable pageable) {
     return scheduleRepository.findAvailableSchedulesByDate(date, pageable)
         .map(ProductScheduleInfoResult::from);
